+++ editor/src/lib/EditingPages/CourseEditor.svelte
@@ -15,7 +15,7 @@
     draggedOver,
     originalEntry = $bindable(),
     activeEntry = $bindable(),
-    entryIsCopy,
+    entryIsNew,
     dropPriority = $bindable(),
     resetList
   }: {
@@ -24,7 +24,7 @@
     draggedEntry: Entry | null;
     originalEntry: string;
     activeEntry: Entry | null;
-    entryIsCopy: boolean;
+    entryIsNew: boolean;
     dropPriority: boolean;
     resetList: Function;
   } = $props();
@@ -32,7 +32,7 @@
   let serverMessage: ServerMessage;
 
   async function handleSubmit() {
-    const method = entryIsCopy
+    const method = entryIsNew
       ? // New chapter
         'POST'
       : // Existing chapter
@@ -58,7 +58,7 @@
   class:dragged-over={draggedOver}
   in:fly={{ y: -15, duration: 600 }}
 >
-  <NewOrEditingLabel entry={course} {entryIsCopy} />
+  <NewOrEditingLabel entry={course} {entryIsNew} />
 
   <!-- TRANSLATIONS -->
   <div class="translation-grid">